Revert file rename
